+++ Frontend/src/components/calendar/Calendar.css
@@ -5,7 +5,6 @@
     background-color: grey;
     margin: 10px;
     padding: 10px;
-    width: 100%;
 }
 
 .fc-media-screen {
