+++ Frontend/src/components/mentee/mentee-dashboard/MenteeDashboard.jsx
@@ -1,6 +1,6 @@
 import { useState} from 'react'
 import './MenteeDashboard.css'
-import Calendar from '../../Calendar'
+import Calendar from '../../calendar/Calendar'
 import Requests from '../../mentor/mentor-requests/Requests'
 import { Link } from 'react-router-dom';
 import MenteeMatches from '../mentee-matches/MenteeMatches'
